Joomla BMI calculator with graph
This is a nice Body Mass Index (BMI) Calculator Module you can use on
your Joomla! 1.5 /1.6/ 1.7/ 2.5 site.

You can even change the language with the language file and change the
layout with the css file.

This module is updated to a newer version !!!
Now it includes a graph and you can adjust the unit selection and sex
selection at the admin part of the module